pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "malice-network"
version = "0.0.1"
description = "Python SDK for Malice Network C2 Framework"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ChainReactors", email = "chainreactors@example.com"}
]
maintainers = [
    {name = "ChainReactors", email = "chainreactors@example.com"}
]
keywords = ["malice", "c2", "red-team", "penetration-testing", "grpc", "async"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking",
    "Framework :: AsyncIO",
]
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "betterproto2>=0.9.0",
    "grpclib>=0.4.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "flake8>=5.0.0",
    "pre-commit>=2.20.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-autodoc-typehints>=1.19.0",
]
stub-generation = [
    "monkeytype>=22.2.0",
    "mypy>=1.0.0",
]
build = [
    "betterproto2_compiler>=0.9.0",
]

[project.urls]
Homepage = "https://github.com/chainreactors/malice-network"
Documentation = "https://chainreactors.github.io/wiki/IoM/"
Repository = "https://github.com/chainreactors/malice-network"
"Bug Tracker" = "https://github.com/chainreactors/malice-network/issues"
Changelog = "https://github.com/chainreactors/malice-network/blob/main/CHANGELOG.md"

[project.scripts]
malice-generate = "IoM.generate:main"

[tool.setuptools]
packages = ["IoM"]
include-package-data = true

[tool.setuptools.package-data]
IoM = ["*.pyi", "proto/**/*.py", "proto/**/*.pyi"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "grpclib.*",
    "betterproto.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["IoM"]
omit = [
    "*/tests/*",
    "*/proto/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]